+++ tailwind.config.js
@@ -2,6 +2,7 @@
 export default {
   content: [
     "./index.html",
+    "./navigator.html",
     "./src/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {